Added tuple<->list conversion
Note on void-like tuple members: Previously in uFunctions
a tuple t2Xli could be converted to li if 'X' has a void value.
That is, tuple members that are void can be 'converted away'.

This leads to an ambiguity with the ability to convert tuples to lists.
How shall we convert 't2Xli' to 'la'?
- According to the old strategy we shall check if 'X' is void and then
  convert the second member of the tuple 'li' to 'la' by wrapping
  all 'i'nteger members to an 'a'ny.
- One can of course convert the members directly, the first 'X' shall be
  the first element in the resulting list (even if it is void) and the
  second 'li' member shall be the second.

Since most people expect the tuple<->list conversion to go from member
to member, we adopt the second behaviour whenever the policy includes
'allow_converting_tuple_list' and the first one when it doesn't.