perf: don't make cloned chunks read-only
The main change is to disallow chunk::unshare() to force read-only.
Doing so no benefits. (After unshare() a chunk uses a writable sview
only if it is the exclusive user of it. If it was a writable sview to
start with, why remove writability? That will only cause a copy when the
chunk is written to.)
Then we have also removed the ability to create an owning sview that is
not writable as it was never used.