uf::error drives from std::logic_error
According to the stdandard std::runtime_errors are "due to events beyond
the scope of the program and can not be easily predicted", whereas
std::logic_errors are "a consequence of faulty logic within the program
such as violating logical preconditions or class invariants and may be
preventable". This makes uf::error more a logic error than a runtime
one.