[wany] fix: set any sizes on set()
When the content of an 'any' is set, we store the length of the type and
value as these cost little. Note: the length of the value of any 'a'
among our parents is still not updated.

This update was deemed necessary so that linear search on maps work,
if the value member of a map is 'a' and we set some elements.