[wany] fix: fully correct handling of any sizes
Before this patch, when setting a child wview's content via wview::set()
we have only checked its parent. If it was an any, we updated the type
and value length. However, we have not updated any further any parent,
resulting in incorrect sizes there making scanning operations (e.g.,
operator[]) invalid.
In addition, when inserting or deleting a value from a map, list,
optional or tuple, we did not update the any sizes in the parents at
all.
(Any sizes mean the type length and value length stored in a serialized
any value.)

This patch fixes these issues.
- At any set operation, we check parents (all the way to the top) and
  if we find any 'any', we adjust its value length.
- This unnecessitates to calculate these any sizes before any flatten()
  call, hence this is removed. (Probably an overall speed improvement.)
- We also update the parent's any sizes on insert and delete.
- We introduced flatten_type_size() as helper.
- Added a test.
- Uplifted the natvis file, to cater for sview::writable being atomic.
- Enhanced the natvis file to show 4 byte long chunks as if they were
  a size value (serialized uint32_t), as well. This makes decoding size
  values (string, map, list sizes) easier.