Move func bodies that use error_value
Clang in C++20 mode insists that it cannot instantiate stuff having or using the incomplete error_value type.
(Both g++ for all standards and Clang for C++17 compiles fine, so whether this is a regression or a fix is not clear.)
For example, Clang claims that expected_with_error ctor would need error_value dtor when moving the argument, which may indeed be
a violation of [vector.overview]/p4 (and g++ deferring it until later in the file as an extension), but the *templated* any_v::get*
is a mistery for me.